+++ packages/article-in-depth/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.53.1](https://github.com/newsuk/times-components/compare/@times-components/article-in-depth@3.53.0...@times-components/article-in-depth@3.53.1) (2022-07-06)
+
+**Note:** Version bump only for package @times-components/article-in-depth
+
+
+
+
+
 # [3.53.0](https://github.com/newsuk/times-components/compare/@times-components/article-in-depth@3.52.23...@times-components/article-in-depth@3.53.0) (2022-07-04)