+++ packages/article-magazine-standard/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.51.1](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-standard@3.51.0...@times-components/article-magazine-standard@3.51.1) (2022-07-06)
+
+**Note:** Version bump only for package @times-components/article-magazine-standard
+
+
+
+
+
 # [3.51.0](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-standard@3.50.23...@times-components/article-magazine-standard@3.51.0) (2022-07-04)